Add sample shell execution test
This was just proof of concept for getting a subprocess mock up and
running. Currently this is going through testfixtures, and seems to be
functioning as desired.

In addition, in order to make the tests work, and in order for the
execute() function to return a string with limited overhead, stdout is
deocded as it's read in shell.execute().